Better citre config, fix org capture, steal doom git gutter
Citre jump is now g j in evil, allowing evil-goto-definition to do it's magic
with g d when citre isn't there or when lsp is available.
Add GNU global config for jumping around java projects, needs special config to
work without issues on Windows.
Create a wrap function for quotes and double quotes, and bind it, C-c w d and
C-c w q or C-c " and C-c '.
Concat doesn't work for org capture, so I fixed the template that used it.