Big completion fixes, evil scroll up binding
Instead of setting a default capf with all of the stuff I want, I've set up
hooks to add all of the capfs I want to the different mode types that I want to
use them in. This means my preferred capfs are never overridden, instead they
are checked first, then whatever capf a specific mode wants to use.

I've also disabled a lot of the tempel stuff, while in most modes it's great,
both the navigation through the snippets and the abbrev mode were messing with
my coding. Will need to watch closely for yasnippet-capf to release on melpa.