Add yas capf to replace tempel in many places, consult tweaks
yasnippet-capf replaces tempel in everything but org and text modes where both
are used.

Configured consult-buffer (C-x b) and consult-notes (C-c s) to have a .5s delay
on preview on Windows. This noticably increases performance.

Rebinds:
C-f :: (evil-normal) Now consult-outline which is a generic way to search files much like
consult-imenu. It is limited to the current file.
C-t :: (evil-normal) Now consult-imenu-multi (previously C-f) I made this change since C-t is
a more neutral binding, and is shared with other code editors (like vscode).
C-s :: Now consult-ripgrep if ripgrep is installed. If not, it is bound to
consult-git-grep which is the original binding.
C-: :: Now embark-collect which is a very powerful command used in the
minibuffer, it exports the content to a separate buffer, making it more
permanent and also easily searched. Very useful with consult-line and other
minibuffer search commands.