Add icon/emoji support cross-platform, marginalia display tweak
Adds all-the-icons, updated readme with instructions. Integrates into vertico,
marginalia, dired, and doom-modeline.
Update marginalia's maximum column width setting so that it will always
calculate the column width dynamically. Previously, any time the window width
was > 160 characters, it would default to 80 max. This is practically all the
time, since in full screen the window width is > 200, which means there was a
lot of dead space in commands like C-h v or C-x b. This fixes that issue.