Remove evil smartparens, g spc to evil join whitespace, treesitter
Font lock level for treesitter sets how many layers of text get colored, higher
the number the more stuff that is colored.
Evil join whitespace is g j in standard vim however it is citre jump in my
config, so I mapped it to g spc so I can retain the useful command.
Removed evil-smartparens because it is not as smart as smartparens at deleting
pairs, also it makes deleting long lines take far too long, and it messes up
yank. Generally, it is also nice that I have a way to cheat strict mode now in
vim when smartparens messes up.