Update elpaca installer script from web, enable smartparens again
An updated elpaca installer was posted last week which causes a warning on
launch after updating. Updated script to remove warning.
When evil-smartparens was removed from the ini, it also removed the code
enabling smartparens, added that back in.

Side note: First package update in a while, everything seems to work fine.