Undo treesit changes, bring back cape-dict hacks, org hook cleanup
See the rant above treesit-auto for details on that.
Minad had a user report a performance issue with my changes in cape-dict and
reverted the change, so now I am back to my old change for now. Ideally, I
should create a new package for this, I'm tired of changing my config for this
issue and it would be a good learning experience in my opinion.
The org hook I removed is unnecessary from before I had a word wrap setup.