Unbind C-c s from consult-notes, / from consult-line and many others
Changed binding of consult-notes to C-c n s for consistency.
Changed binding of consult-line to C-c s, allowing / to be the default evil
binding. Changed consult-multi-line to C-c C-s for consistency.
Changed consult-outline to C-c t o since it is rarely used.
Set C-f in normal mode to evil-avy-goto-word-or-subword-1. This allows jumping
to a word or subword in any active buffer.
Add flush-empty-lines function to delete all empty lines in a file.
Add hiccup-cli to readme and add bindings. Overrides html-to-hiccup bindings
when hiccup-cli is present.