Add ace-window mappings to insert mode C-w, C-S-a sp-begin, more
Bind sp-beginning-of-sexp to C-S-a to match sp-end-of-sexp.
Bind ace-window to C-w in insert mode, and C-S-w to ace-swap-window. This is so
that I can keep certain buffers that would prefer insert mode in that mode.
Add auto lsp to clojure, I will probably never use it on windows and it needs
it.
Alter hiccup-cli setup, uses executable-find which only works if running emacs
from the terminal in macos but the package check doesn't work anymore, or maybe
it never did? May need to revisit exwm after this.
Change reference to clojure-ts-mode to clojure-mode.