Finish configuring AI stuff, minor doom-modeline change to match doc
In gptel, I added some basic global mappings, and added S-RET as a gptel-send
option.
In copilot-chat, I added a global mapping and fixed a bug where it was causing
issues with doom modeline.
For doom-modeline, the doc calls doom-modeline in the init secton so I changed
mine to match.