Add some configuration for which-func to avoid crashes, add semantic
Which function mode was causing crashes in random modes that would have not been
useful anyways such as sql-mode, so I've started a list of useful and non-crashy
modes to whitelist for it.

Semantic mode is basically a breadcrumb, seems like it has a lot of promise, but
need to be careful about it because it's a bit heavy.