Local AI setup, more sane ai bindings, rebind org-agenda
All ai functions have been moved from C-c C-a to C-c a.
Org agenda was originally on C-c a, now it is C-c a o.
Moved org-store-link to C-c n S-l (previously C-c l). This is because C-c l is
rebound by lsp-mode.
Rewrote the ai setup functions, now it is easier to set up different ai agents
or defaults for different situations or installs.