Bind C-k and evil-lookup to lsp-describe-thing-at-point, cleanup
lsp-describe-thing-at-point bindings:
This setup makes sure C-k is available even in insert to lookup the thing at
point, the evil-lookup binding also makes sure outside of clojure K will open up
the lsp documentation in normal mode. (in clojure, it has a special K binding
that is actually better than the lsp binding)

Also did an auto-format on the lsp-booster code, it had some weird things going
on there.

In ejc-sql, added ; as the separator so C-c C-c only runs the sql statement
under the cursor and not the whole buffer.