Native treesit plugin cutover done.
Tested the treesit setup with windows and it works! Will likely need a lot of
changes to hooks and so on so that lsp and other common language hooks start up.

Changed html files to use plain html-mode by default since it is less buggy.

Add clojure-ts-mode back (treesit migration).

Add some documentation for budget calc functions.

Clean up which-func-modes, adding js back in because of treesit.