Add dumb jump, perl config, fix elpaca issues on windows, rm org-journal
dumb-jump:
Quick and dirty xref function when not wanting to mess with TAGS or LSP.
Integrated with the usual definition functions (M-., gd) and has a complex
hydra version bound to C-c j (previously org-journal).

perl:
Initial configuration based on stuff from the internet.

elpaca:
Fix a hanging issue on windows due to the elpaca-queue-limit being too low.

org-journal:
I don't use it and wanted the C-c j binding to use for dumb-jump, so it has been removed.