Fix vdiff and gnus, move ideas to specific org file
vdiff: Was not loading hydra extensions and bind does not work for the prefix
map, moved to under hydra to hopefully help with load order problems in Windows.

gnus: Map does not load when adding bindings, so the binding was not applied,
the old mapping binds redirects to the new binding anyways, so I removed the
config rather than fixing.

org: Existing ideas file was not used and I don't want to mix it with side
project stuff.