fix: Make install less buggy, remove warnings.
init.el:
Elpaca will get a memory leak on linux if you use elpaca-wait on the first
startup. To offset this issue, added a file to the elpa folder which triggers
use of elpaca-wait.

early-init.el:
Master emacs has a warning on startup if multiple headers are in the early-init,
so I removed the byte-compile one.

corfu.el:
Corfu terminal throws a warning since it has become irrelevant with newer
versions of emacs.