fix: Missing elpa dir, clojure-ts-mode has a major bug
init.el:
The elpa directory doesn't exist at the point that this code is ran on linux, so
make it just in case.

lang.el:
Clojure's treesit package works great when you're not using cider. This may be
worthy of a bug report, but for now I'm just moving back to the old
clojure-mode.