fix: LSP issues, testing removal of lsp-booster (req full plugin reinstall)
Removal of LSP booster:
This requires a full plugin re-install to change LSP to not use plists.
This change is because lsp-booster is having problems on Windows, and the
somewhat recent statistics showing that plists are actually slower in many cases
than the built in JSON parser included in emacs 30+. This repo does not try to
support anything less than the latest stable version, so assuming nothing crazy
happens in the coming months all data for lsp-booster should be purged (exes,
elisp, folders).

Fixed weird issue with the completion popup:
completion-category settings are apparently necessary for this, so I added it to
the mode hook.

Adding mason (commented out for now):
It would be nice to run the latest stable releases of jdtls especially, however
it seems that this plugin is not well suited to Windows, or maybe it is the lsp
booster advice that messed it up? In any case I do want to spend more time
getting this working which is why it's partially included now.